+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add templates for listing offers
 - Add moderation for the offers
 - add offer edit form
+- template of sharing collection request #831
 
 ### Changed
 - update all Javascript Dependencies